GRHB-506: * move course card component to common components folder
