+++ mobile/src/components/course-card/common/maps/maps.ts
