+++ mobile/src/components/course-card/common/maps/vendor-key-to-image.map.ts
