+++ mobile/src/components/course-card/components/components.ts
