+++ mobile/src/components/course-card/components/vendor-label/style.ts
