+++ mobile/src/components/course-card/styles.ts
