+++ mobile/src/components/courses/components/components.ts
@@ -1,2 +1,1 @@
 export { AddCourse } from './add-course/add-course';
-export { CourseCard } from './course-card/course-card';
